requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.25.0
diffusers>=0.21.0
accelerate>=0.24.0
peft>=0.6.0
Flask>=2.3.0
flask-cors>=4.0.0
clip-by-openai
huggingface_hub>=0.17.0
Pillow>=9.0.0
numpy>=1.24.0
tqdm>=4.64.0
packaging>=21.0
tensorboard>=2.12.0
